// MyMusic.Client/src/model/soundalikeDuplicateGroup.ts
// @ts-nocheck
/**
Generated by orval 🍺
Do not edit manually.
MyMusic.Server | v1
OpenAPI spec version: 1.0.0
*/import type { SoundalikeDuplicateGroupSecondaryActions } from './soundalikeDuplicateGroupSecondaryActions';
import type { SoundalikeSongItem } from './soundalikeSongItem';

export interface SoundalikeDuplicateGroup {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  nonConformityId: number;
  createdAt: string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$ */
  matchScore: number;
  songs: SoundalikeSongItem[];
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  primarySongId?: number | null;
  secondaryActions?: SoundalikeDuplicateGroupSecondaryActions;
}