// MyMusic.Client/src/model/soundalikeSongItem.ts
// @ts-nocheck
/**
Generated by orval 🍺
Do not edit manually.
MyMusic.Server | v1
OpenAPI spec version: 1.0.0
*/ import type { ListSongsAlbum } from "./listSongsAlbum";
import type { ListSongsArtist } from "./listSongsArtist";
import type { ListSongsGenre } from "./listSongsGenre";

export interface SoundalikeSongItem {
	/** @pattern ^-?(?:0|[1-9]\d*)$ */
	id: number;
	/**
	 * @nullable
	 * @pattern ^-?(?:0|[1-9]\d*)$
	 */
	cover: number | null;
	/**
	 * @nullable
	 * @pattern ^-?(?:0|[1-9]\d*)$
	 */
	coverWidth: number | null;
	/**
	 * @nullable
	 * @pattern ^-?(?:0|[1-9]\d*)$
	 */
	coverHeight: number | null;
	title: string;
	artists: ListSongsArtist[];
	album: ListSongsAlbum;
	genres: ListSongsGenre[];
	/**
	 * @nullable
	 * @pattern ^-?(?:0|[1-9]\d*)$
	 */
	year: number | null;
	duration: string;
	/**
	 * @nullable
	 * @pattern ^-?(?:0|[1-9]\d*)$
	 */
	bitrate: number | null;
	/** @pattern ^-?(?:0|[1-9]\d*)$ */
	size: number;
	hasLyrics: boolean;
	isExplicit: boolean;
	createdAt: string;
	/** @nullable */
	addedAt?: string | null;
}